Move Price components to separated file
