+++ src/custom/pages/Swap/styled.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components/macro'
+import { RowBetween } from 'components/Row'
+
+export const LowerSectionWrapper = styled(RowBetween).attrs((props) => ({
+  ...props,
+  align: 'center',
+  flexDirection: 'row',
+  flexWrap: 'wrap',
+  minHeight: 24,
+}))`
+  > .price-container {
+    display: flex;
+    gap: 5px;
+  }
+`
